.NET MVC project for Medical History
Hello :=)

Info: visual studio 2015 MVC project using .NET 4.6.2 and Entity
Framwork version 6.0.0.0.

Description:

This is currently an ongoing project I have with my mother. She needs a
medical history CRUD application, with one exception,
no deleting. This is because she wants to keep a long term history
record of all the information. The model is straight forward,
and the UI will provide access to lists, data entry, and reports. Some
other future plans are as follows:
(1) Create a phone app that will, for each report, synch/download latest
report data in JSON using a .NET MVC Web API REST entry point.
(2) Create a UI feature that can push doctor appointments to her outlook
using "Outlook Push Notifications REST API".

One important note is that this application is not public. It will
reside on a home intranet and she will be the only user.

There is a backup of the database in the folder Database.

------ Taruchu :=)